+++ powerauth-java-cmd-lib/src/main/java/com/wultra/security/powerauth/lib/cmd/steps/model/FetchConfigStepModel.java
@@ -1,6 +1,6 @@
 /*
  * PowerAuth Command-line utility
- * Copyright 2018 Wultra s.r.o.
+ * Copyright 2026 Wultra s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,11 +16,9 @@
  */
 package com.wultra.security.powerauth.lib.cmd.steps.model;
 
-
 import com.wultra.security.powerauth.crypto.lib.v4.model.context.SharedSecretAlgorithm;
 import com.wultra.security.powerauth.lib.cmd.steps.model.data.EncryptionHeaderData;
 import com.wultra.security.powerauth.lib.cmd.steps.model.data.MasterPublicKeyData;
-import com.wultra.security.powerauth.lib.cmd.steps.model.feature.DryRunCapable;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
@@ -28,19 +26,15 @@ import java.security.PublicKey;
 import java.util.Map;
 
 /**
- * Model representing parameters of the step for sending encrypted data to intermediate server.
+ * Model representing parameters of the step for fetching the secure configuration over end-to-end
+ * encryption (V4 only).
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class EncryptStepModel extends BaseStepModel
-        implements EncryptionHeaderData, DryRunCapable, MasterPublicKeyData {
-    /**
-     * Request data.
-     */
-    private byte[] data;
-
+public class FetchConfigStepModel extends BaseStepModel
+        implements EncryptionHeaderData, MasterPublicKeyData {
     /**
      * Application key.
      */
@@ -50,11 +44,6 @@ public class EncryptStepModel extends BaseStepModel
      * Application secret.
      */
     private String applicationSecret;
-
-    /**
-     * Flag indicating that this step should be terminated before the networking call.
-     */
-    private boolean dryRun;
 
     /**
      * Master Server Public Key for P-256, a value specific for given application.
@@ -82,17 +71,15 @@ public class EncryptStepModel extends BaseStepModel
     private SharedSecretAlgorithm sharedSecretAlgorithm;
 
     /**
-     * ECIES encryption scope.
+     * Encryption scope: {@code application} or {@code activation}.
      */
     private String scope;
 
     @Override
     public Map<String, Object> toMap() {
         Map<String, Object> context = super.toMap();
-        context.put("DATA", data);
         context.put("APPLICATION_KEY", applicationKey);
         context.put("APPLICATION_SECRET", applicationSecret);
-        context.put("DRY_RUN", dryRun);
         context.put("MASTER_PUBLIC_KEY_P256", masterPublicKeyP256);
         context.put("MASTER_PUBLIC_KEY_P384", masterPublicKeyP384);
         context.put("MASTER_PUBLIC_KEY_MLDSA65", masterPublicKeyMlDsa65);
@@ -105,10 +92,8 @@ public class EncryptStepModel extends BaseStepModel
     @Override
     public void fromMap(Map<String, Object> context) {
         super.fromMap(context);
-        setData((byte[]) context.get("DATA"));
         setApplicationKey((String) context.get("APPLICATION_KEY"));
         setApplicationSecret((String) context.get("APPLICATION_SECRET"));
-        setDryRun((boolean) context.get("DRY_RUN"));
         setMasterPublicKeyP256((PublicKey) context.get("MASTER_PUBLIC_KEY_P256"));
         setMasterPublicKeyP384((PublicKey) context.get("MASTER_PUBLIC_KEY_P384"));
         setMasterPublicKeyMlDsa65((PublicKey) context.get("MASTER_PUBLIC_KEY_MLDSA65"));
@@ -118,3 +103,4 @@ public class EncryptStepModel extends BaseStepModel
     }
 
 }
+